+++ eggs/Naaya/trunk/Products/NaayaCore/managers/import_export.py
@@ -5,16 +5,8 @@ import csv, codecs
 from StringIO import StringIO
 import urllib
 import simplejson as json
-try:
-    import xlwt
-    excel_export_available = True
-except:
-    excel_export_available = False
-try:
-    import xlrd
-    excel_read_available = True
-except:
-    excel_read_available = False
+import xlwt
+import xlrd
 
 import transaction
 from Acquisition import Implicit
@@ -68,7 +60,6 @@ class CSVImportTool(Implicit, Item):
             filename = schema.title_or_id() + ' bulk upload.csv'
 
         elif file_type == 'Excel':
-            assert excel_export_available
             ret = generate_excel(columns, [[]])
             content_type = 'application/vnd.ms-excel'
             filename = schema.title_or_id() + ' bulk upload.xls'
@@ -134,7 +125,6 @@ class CSVImportTool(Implicit, Item):
                 }
 
         if file_type == 'Excel':
-            assert excel_export_available
             try:
                 wb = xlrd.open_workbook(file_contents=data.read())
                 ws = wb.sheets()[0]
@@ -359,7 +349,6 @@ class ExportTool(Implicit, Item):
             filename = '%s Export.csv' % meta_type
 
         elif file_type == 'Excel':
-            assert excel_export_available
             ret = self.generate_excel_output(meta_type, objects)
             content_type = 'application/vnd.ms-excel'
             filename = '%s Export.xls' % meta_type